chore: move testdata into generated folder
