+++ apps/watch/src/components/AudioDialog/AudioLanguageDialog.stories.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { Meta, Story } from '@storybook/react'
 import { watchConfig } from '../../libs/storybook'
 import { VideoProvider } from '../../libs/videoContext'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { AudioLanguageDialog } from '.'
 
 const AudioLanguageDialogStory = {
