+++ apps/watch/src/components/HomePage/HomePage.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 import { watchConfig } from '../../libs/storybook'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { HomePage } from '.'
 
 const HomePageStory: ComponentMeta<typeof HomePage> = {
