+++ apps/watch/src/components/ShareDialog/ShareDialog.stories.tsx
@@ -8,7 +8,7 @@ import {
 } from '../../../__generated__/VideoContentFields'
 import { watchConfig } from '../../libs/storybook'
 import { VideoProvider } from '../../libs/videoContext'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { ShareDialog } from './ShareDialog'
 
 const ShareDialogStory = {
