+++ apps/watch/src/components/VideoCard/VideoCard.spec.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { VideoCard } from '.'
 
 describe('VideoCard', () => {
