+++ apps/watch/src/components/VideoCard/VideoCard.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 import Grid from '@mui/material/Grid'
 import { watchConfig } from '../../libs/storybook'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { VideoCard } from '.'
 
 const VideoCardStory: ComponentMeta<typeof VideoCard> = {
