+++ apps/watch/src/components/VideoContainerPage/VideoContainerPage.stories.tsx
@@ -2,7 +2,7 @@ import { ComponentMeta, ComponentStory } from '@storybook/react'
 import { MockedProvider } from '@apollo/client/testing'
 import { watchConfig } from '../../libs/storybook'
 import { VideoProvider } from '../../libs/videoContext'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { GET_VIDEO_CHILDREN } from './VideoContainerPage'
 import { VideoContainerPage } from '.'
