+++ apps/watch/src/components/VideoContentPage/AudioLanguageButton/AudioLanguageButton.spec.tsx
@@ -1,6 +1,6 @@
 import { render, fireEvent, waitFor } from '@testing-library/react'
 import { VideoProvider } from '../../../libs/videoContext'
-import { videos } from '../../Videos/testData'
+import { videos } from '../../Videos/__generated__/testData'
 import { AudioLanguageButton } from '.'
 
 describe('AudioLanguageButton', () => {
