+++ apps/watch/src/components/VideoContentPage/VideoContentPage.stories.tsx
@@ -2,7 +2,7 @@ import { ComponentProps } from 'react'
 import { Story, Meta } from '@storybook/react'
 import { watchConfig } from '../../libs/storybook'
 import { VideoProvider } from '../../libs/videoContext'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { VideoContentPage } from '.'
 
 const VideoContentPageStory = {
