+++ apps/watch/src/components/VideoContentPage/VideoHero/VideoHero.stories.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import { ThemeName, ThemeMode } from '@core/shared/ui/themes'
 import { watchConfig } from '../../../libs/storybook'
 import { VideoProvider } from '../../../libs/videoContext'
-import { videos } from '../../Videos/testData'
+import { videos } from '../../Videos/__generated__/testData'
 import { VideoHero } from './VideoHero'
 
 const VideoHeroStory = {
