+++ apps/watch/src/components/VideoGrid/VideoGrid.spec.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render } from '@testing-library/react'
-import { videos } from '../Videos/testData'
+import { videos } from '../Videos/__generated__/testData'
 import { VideoGrid } from './VideoGrid'
 
 describe('VideoGrid', () => {
